+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ProblemProvider can now detect unquoted dates in parameter values [#1197](https://github.com/HicServices/RDMP/issues/1197)
 - Added a `NLog.template.config` file to releases for easily enabling diagnostics logging to disk (NLog logging is still disabled by default for the windows client)
 - Performance metrics (refresh time, menu building times) are now passed to NLog logging when enabled in windows client [#1227](https://github.com/HicServices/RDMP/issues/1227)
+- Plugin UploadFileUseCase pipeline components can now declare `IPipelineRequirement<IBasicActivateItems>`
 
 ### Changed